Ajuste a convenios para homologar con CU 43

+++ src/main/java/com/imss/sivimss/planfunerario/model/request/ReporteDto.java
@@ -13,7 +13,7 @@ public class ReporteDto {
 	private String tipoReporte;
 	private String rutaNombreReporte;
 	private String folio;
-	private Integer costoRenovacion;
+	private Double costoRenovacion;
 	private Integer idConvenio;
 	private String observaciones;
 	private String directoraFibeso;
